# backend/gateway/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /collection:
        get:
            tags:
                - CollectionService
            description: 列出用户的收藏夹
            operationId: CollectionService_ListCollection
            parameters:
                - name: pagination.page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pagination.size
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.v1.ListCollectionResponse'
        put:
            tags:
                - CollectionService
            description: 更新收藏夹信息
            operationId: CollectionService_UpdateCollection
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/gateway.v1.UpdateCollectionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.v1.UpdateCollectionResponse'
        post:
            tags:
                - CollectionService
            description: 创建收藏夹
            operationId: CollectionService_CreateCollection
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/gateway.v1.CreateCollectionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.v1.CreateCollectionResponse'
        delete:
            tags:
                - CollectionService
            description: 删除收藏夹
            operationId: CollectionService_RemoveCollection
            parameters:
                - name: id
                  in: query
                  description: '@gotags: json:"id,omitempty,string"'
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.v1.RemoveCollectionResponse'
    /collection/video:
        get:
            tags:
                - CollectionService
            description: 列出收藏夹中的视频
            operationId: CollectionService_ListVideo4Collection
            parameters:
                - name: collectionId
                  in: query
                  description: '@gotags: json:"collectionId,omitempty,string"'
                  schema:
                    type: string
                - name: pagination.page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pagination.size
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.v1.ListVideo4CollectionResponse'
        post:
            tags:
                - CollectionService
            description: 将视频添加到收藏夹
            operationId: CollectionService_AddVideo2Collection
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/gateway.v1.AddVideo2CollectionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.v1.AddVideo2CollectionResponse'
        delete:
            tags:
                - CollectionService
            description: 从收藏夹中移除视频
            operationId: CollectionService_RemoveVideoFromCollection
            parameters:
                - name: collectionId
                  in: query
                  description: '@gotags: json:"collectionId,omitempty,string"'
                  schema:
                    type: string
                - name: videoId
                  in: query
                  description: '@gotags: json:"videoId,omitempty,string"'
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/gateway.v1.RemoveVideoFromCollectionResponse'
...
components:
    schemas:
...
        gateway.v1.AddVideo2CollectionRequest:
            type: object
            properties:
                collectionId:
                    type: string
                    description: '@gotags: json:"collectionId,omitempty,string"'
                videoId:
                    type: string
                    description: '@gotags: json:"videoId,omitempty,string"'
        gateway.v1.AddVideo2CollectionResponse:
            type: object
            properties: {}
...
        gateway.v1.Collection:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                userId:
                    type: string
                    description: '@gotags: json:"userId,omitempty,string"'
                name:
                    type: string
                description:
                    type: string
...
        gateway.v1.CreateCollectionRequest:
            type: object
            properties:
                name:
                    type: string
                description:
                    type: string
        gateway.v1.CreateCollectionResponse:
            type: object
            properties:
                collection:
                    $ref: '#/components/schemas/gateway.v1.Collection'
...
        gateway.v1.ListCollectionResponse:
            type: object
            properties:
                collections:
                    type: array
                    items:
                        $ref: '#/components/schemas/gateway.v1.Collection'
                pagination:
                    $ref: '#/components/schemas/gateway.v1.PaginationResponse'
...
        gateway.v1.ListVideo4CollectionResponse:
            type: object
            properties:
                videos:
                    type: array
                    items:
                        $ref: '#/components/schemas/gateway.v1.Video'
                pagination:
                    $ref: '#/components/schemas/gateway.v1.PaginationResponse'
...
        gateway.v1.PaginationResponse:
            type: object
            properties:
                page:
                    type: integer
                    format: int32
                total:
                    type: integer
                    format: int32
                count:
                    type: integer
                    format: int32
...
        gateway.v1.RemoveCollectionResponse:
            type: object
            properties: {}
...
        gateway.v1.RemoveVideoFromCollectionResponse:
            type: object
            properties: {}
...
        gateway.v1.UpdateCollectionRequest:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                name:
                    type: string
                description:
                    type: string
        gateway.v1.UpdateCollectionResponse:
            type: object
            properties:
                collection:
                    $ref: '#/components/schemas/gateway.v1.Collection'
...
        gateway.v1.Video:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                author:
                    $ref: '#/components/schemas/gateway.v1.VideoAuthor'
                playUrl:
                    type: string
                coverUrl:
                    type: string
                favoriteCount:
                    type: string
                    description: '@gotags: json:"favoriteCount,omitempty,string"'
                commentCount:
                    type: string
                    description: '@gotags: json:"commentCount,omitempty,string"'
                isFavorite:
                    type: boolean
                title:
                    type: string
                isCollected:
                    type: boolean
                collectedCount:
                    type: string
                    description: '@gotags: json:"collectedCount,omitempty,string"'
        gateway.v1.VideoAuthor:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                name:
                    type: string
                avatar:
                    type: string
                isFollowing:
                    type: boolean
...